Job - create

+++ src/main/kotlin/com/ex2/jobs/auth/repo/RoleRepository.kt
@@ -3,4 +3,4 @@ package com.ex2.jobs.auth.repo
 import com.ex2.jobs.auth.entities.RoleEntity
 import org.springframework.data.jpa.repository.JpaRepository
 
-interface RoleRepository : JpaRepository<RoleEntity, Int>
+interface RoleRepository : JpaRepository<RoleEntity, Long>
